requirements.txt
Flask==2.3.3
flask-cors==4.0.0
llama-cpp-python==0.2.11
requests==2.31.0